split http client out of the wasm connector client
  This serves two purposes:
  1. It allows reusing the http client bits to implement more clients (e.g. an adaptor client) for wasm
  2. It makes obvious that the connector logic in that module is really just a wrapper around an http client, so we should be able to merge this entirely with a single generic adaptor client, that can swap its http client on-the-fly depending on the target environment!
