+++ rust/crates/cardano-connect-wasm/src/lib.rs
@@ -1,3 +1,6 @@
+mod http_client;
+pub use http_client::HttpClient;
+
 mod connector;
 pub use connector::CardanoConnector;
